Refactor: page js 이동
