+++ careerForDeveloper/src/main/webapp/reactjs/src/components/App.js
@@ -8,10 +8,10 @@ import {BrowserRouter as Router, Route, Routes} from "react-router-dom";
 import Main from "./Main";
 import Login from "./Login";
 import Logout from "./Logout";
-import Signup from "./Signup";
+import Signup from "../pages/Signup";
 import DropUser from "./DropUser";
 import Update from "./Update";
-import Posts from "./Posts";
+import Posts from "../pages/Posts";
 import PostView from "./PostView";
 import NewPost from "./NewPost";
 import DeletePost from "./DeletePost";
@@ -20,14 +20,13 @@ import UpdateComment from "./UpdateComment";
 import UpdatePost from "./UpdatePost";
 import DeleteReply from "./DeleteReply";
 import NewProject from "./NewProject";
-import ProjectView from "./ProjectView";
+import ProjectView from "../pages/ProjectView";
 import DeleteProject from "./DeleteProject";
 import UpdateProject from "./UpdateProject";
-import Projects from "./Projects";
+import Projects from "../pages/Projects";
 import ProjectRequest from "./ProjectRequest";
 import ProjectRequestView from "./ProjectRequestView";
-import RequestView from "./RequestView";
-import UserPage from "../pages/UserPage";
+import RequestView from "../pages/RequestView";
 import MyPage from "../pages/MyPage";
 import UserRequest from "../pages/UserRequest";
