+++ careerForDeveloper/src/main/webapp/reactjs/src/components/iconify/index.js
@@ -1,0 +1,1 @@
+export { default } from './Iconify';
