+++ careerForDeveloper/src/main/webapp/reactjs/src/pages/Signup.js
@@ -1,0 +1,322 @@
+import React from "react";
+import axios from "axios";
+import EmailValidation from "../components/EmailValidation";
+import '../App.css';
+import {useNavigate} from "react-router-dom";
+import Avatar from '@mui/material/Avatar';
+import Button from '@mui/material/Button';
+import CssBaseline from '@mui/material/CssBaseline';
+import TextField from '@mui/material/TextField';
+import Link from '@mui/material/Link';
+import Grid from '@mui/material/Grid';
+import Box from '@mui/material/Box';
+import LockOutlinedIcon from '@mui/icons-material/LockOutlined';
+import Typography from '@mui/material/Typography';
+import Container from '@mui/material/Container';
+import { createTheme, ThemeProvider } from '@mui/material/styles';
+import NavBar from "../components/NavBar";
+
+function Copyright(props) {
+    return (
+        <Typography variant="body2" color="text.secondary" align="center" {...props}>
+            {'Copyright © '}
+            <Link color="inherit" href="https://mui.com/">
+                Your Website
+            </Link>{' '}
+            {new Date().getFullYear()}
+            {'.'}
+        </Typography>
+    );
+}
+
+const theme = createTheme();
+
+// export default function SignUp() {
+//     const handleSubmit = (event) => {
+//         event.preventDefault();
+//         const data = new FormData(event.currentTarget);
+//         console.log({
+//             email: data.get('email'),
+//             password: data.get('password'),
+//         });
+//     };
+//
+//     return (
+//         <ThemeProvider theme={theme}>
+//             <Container component="main" maxWidth="xs">
+//                 <CssBaseline />
+//                 <Box
+//                     sx={{
+//                         marginTop: 8,
+//                         display: 'flex',
+//                         flexDirection: 'column',
+//                         alignItems: 'center',
+//                     }}
+//                 >
+//                     <Avatar sx={{ m: 1, bgcolor: 'secondary.main' }}>
+//                         <LockOutlinedIcon />
+//                     </Avatar>
+//                     <Typography component="h1" variant="h5">
+//                         Sign up
+//                     </Typography>
+//                     <Box component="form" noValidate onSubmit={handleSubmit} sx={{ mt: 3 }}>
+//                         <Grid container spacing={2}>
+//                             <Grid item xs={12} sm={6}>
+//                                 <TextField
+//                                     autoComplete="given-name"
+//                                     name="firstName"
+//                                     required
+//                                     fullWidth
+//                                     id="firstName"
+//                                     label="First Name"
+//                                     autoFocus
+//                                 />
+//                             </Grid>
+//                             <Grid item xs={12} sm={6}>
+//                                 <TextField
+//                                     required
+//                                     fullWidth
+//                                     id="lastName"
+//                                     label="Last Name"
+//                                     name="lastName"
+//                                     autoComplete="family-name"
+//                                 />
+//                             </Grid>
+//                             <Grid item xs={12}>
+//                                 <TextField
+//                                     required
+//                                     fullWidth
+//                                     id="email"
+//                                     label="Email Address"
+//                                     name="email"
+//                                     autoComplete="email"
+//                                 />
+//                             </Grid>
+//                             <Grid item xs={12}>
+//                                 <TextField
+//                                     required
+//                                     fullWidth
+//                                     name="password"
+//                                     label="Password"
+//                                     type="password"
+//                                     id="password"
+//                                     autoComplete="new-password"
+//                                 />
+//                             </Grid>
+//                             <Grid item xs={12}>
+//                                 <FormControlLabel
+//                                     control={<Checkbox value="allowExtraEmails" color="primary" />}
+//                                     label="I want to receive inspiration, marketing promotions and updates via email."
+//                                 />
+//                             </Grid>
+//                         </Grid>
+//                         <Button
+//                             type="submit"
+//                             fullWidth
+//                             variant="contained"
+//                             sx={{ mt: 3, mb: 2 }}
+//                         >
+//                             Sign Up
+//                         </Button>
+//                         <Grid container justifyContent="flex-end">
+//                             <Grid item>
+//                                 <Link href="#" variant="body2">
+//                                     Already have an account? Sign in
+//                                 </Link>
+//                             </Grid>
+//                         </Grid>
+//                     </Box>
+//                 </Box>
+//                 <Copyright sx={{ mt: 5 }} />
+//             </Container>
+//         </ThemeProvider>
+//     );
+// }
+
+function Signup() {
+
+    const [email, setEmail] = React.useState("");
+    const [password, setPassword] = React.useState("");
+    const [passwordChk, setPasswordChk] = React.useState("");
+    const [nickname, setNickname] = React.useState("");
+    const [show, setShow] = React.useState(false);
+    const [certified, setCertified] = React.useState("");
+    const [auth, setAuth] = React.useState(false);
+    const navigate = useNavigate();
+
+    const isEmail = (email) => {
+        const emailRegex =
+            /^(([^<>()\[\].,;:\s@"]+(\.[^<>()\[\].,;:\s@"]+)*)|(".+"))@(([^<>()[\].,;:\s@"]+\.)+[^<>()[\].,;:\s@"]{2,})$/i;
+
+        return emailRegex.test(email);
+    };
+
+    const getEmailAuth = (auth) => {
+        setAuth(auth);
+        if (auth) {
+            setShow(false);
+        }
+    }
+
+    const sendEmail = (e) => {
+      e.preventDefault();
+      if (!email) {
+          return alert("이메일을 입력해주세요");
+      }
+        axios.get("/auth/email", {
+            params: {
+                email: email
+            }
+        })
+            .then((res) => {
+                console.log(res);
+                if (res.data.isSuccess) {
+                    setCertified(res.data.result);
+                    alert("이메일을 전송했습니다.");
+                    setShow(true);
+                }
+                else {
+                    alert("이메일은 전송하지 못했습니다. \n다시 시도해주세요.");
+                }
+            });
+    };
+
+
+
+    const SignupFunc = (e) => {
+        e.preventDefault();
+        if (!email || !password || !passwordChk || !nickname) {
+            return alert("모든 항목을 입력해주세요");
+        }
+        else if (!isEmail(email)) {
+            return alert("ID가 이메일 형식이 아닙니다.");
+        }
+        else if (!auth) {
+            return alert("이메일 인증을 완료해주세요.");
+        }
+        else if (password != passwordChk) {
+            return alert("비밀번호와 비밀번호 확인이 다릅니다.");
+        }
+
+        axios.post("/users",{
+                email: email,
+                pwd: password,
+                nickname: nickname
+            }
+        )
+            .then((res) => {
+                console.log(res);
+                if (res.data.isSuccess) {
+                    alert("회원가입 되었습니다!")
+                    navigate("/login");
+                }
+                else {
+                    alert(res.data.message)
+                }
+            });
+    };
+
+    return (<>
+        <NavBar />
+        <ThemeProvider theme={theme}>
+            <Container component="main" maxWidth="xs">
+                <CssBaseline />
+                <Box
+                    sx={{
+                        marginTop: 8,
+                        display: 'flex',
+                        flexDirection: 'column',
+                        alignItems: 'center',
+                    }}
+                >
+                    <Avatar sx={{ m: 1, bgcolor: 'secondary.main' }}>
+                        <LockOutlinedIcon />
+                    </Avatar>
+                    <Typography component="h1" variant="h5">
+                        Sign up
+                    </Typography>
+                    <Box component="form" noValidate onSubmit={SignupFunc} sx={{ mt: 3 }}>
+                        <Grid container spacing={2}>
+                            <Grid item xs={12} sm={9}>
+                                <TextField
+                                    required
+                                    fullWidth
+                                    id="email"
+                                    label="이메일"
+                                    name="email"
+                                    autoComplete="email"
+                                    onChange={(e) => setEmail(e.target.value)}
+                                />
+                            </Grid>
+                            <Grid item xs={12} sm={3}>
+                                <Button
+                                    onClick={sendEmail}
+                                    fullWidth
+                                    variant="contained"
+                                >
+                                    인증
+                                </Button>
+                            </Grid>
+                            {show && <EmailValidation certified={certified} sendEmailAuth={getEmailAuth}/>}
+                            <Grid item xs={12}>
+                                <TextField
+                                    autoComplete="닉네임을 입력하세요."
+                                    name="nickname"
+                                    required
+                                    fullWidth
+                                    id="nickname"
+                                    label="닉네임"
+                                    autoFocus
+                                    onChange={(e) => setNickname(e.target.value)}
+                                />
+                            </Grid>
+                            <Grid item xs={12}>
+                                <TextField
+                                    required
+                                    fullWidth
+                                    name="password"
+                                    label="비밀번호"
+                                    type="password"
+                                    id="password"
+                                    autoComplete="비밀번호"
+                                    onChange={(e) => setPassword(e.target.value)}
+                                />
+                            </Grid>
+                            <Grid item xs={12}>
+                                <TextField
+                                    required
+                                    fullWidth
+                                    name="passwordChk"
+                                    label="비밀번호 확인"
+                                    type="password"
+                                    id="passwordChk"
+                                    autoComplete="비밀번호 확인"
+                                    onChange={(e) => setPasswordChk(e.target.value)}
+                                />
+                            </Grid>
+                        </Grid>
+                        <Button
+                            type="submit"
+                            fullWidth
+                            variant="contained"
+                            sx={{ mt: 3, mb: 2 }}
+                        >
+                            회원가입
+                        </Button>
+                        <Grid container justifyContent="flex-end">
+                            <Grid item>
+                                <Link href="/login" variant="body2">
+                                    이미 계정이 있으신가요? 로그인
+                                </Link>
+                            </Grid>
+                        </Grid>
+                    </Box>
+                </Box>
+                <Copyright sx={{ mt: 5 }} />
+            </Container>
+        </ThemeProvider>
+        </>
+    );
+}
+
+export default Signup
